synthetics - organize uptime and synthetics e2e tests
