+++ x-pack/plugins/synthetics/e2e/journeys/index.ts
@@ -5,17 +5,5 @@
  * 2.0.
  */
 
-export * from './data_view_permissions';
-export * from './read_only_user';
+export * from './uptime';
 export * from './synthetics';
-export * from './alerts';
-export * from './uptime.journey';
-export * from './step_duration.journey';
-export * from './monitor_details.journey';
-export * from './monitor_name.journey';
-export * from './monitor_management.journey';
-export * from './monitor_management_enablement.journey';
-export * from './monitor_details';
-export * from './locations';
-export * from './private_locations';
-export * from './detail_flyout';
