+++ x-pack/plugins/synthetics/e2e/journeys/synthetics/add_monitor.journey.ts
@@ -7,7 +7,7 @@
 import uuid from 'uuid';
 import { journey, step, expect, Page } from '@elastic/synthetics';
 import { FormMonitorType } from '../../../common/runtime_types';
-import { syntheticsAppPageProvider } from '../../page_objects/synthetics_app';
+import { syntheticsAppPageProvider } from '../../page_objects/synthetics/synthetics_app';
 
 const customLocation = process.env.SYNTHETICS_TEST_LOCATION;
