+++ x-pack/plugins/synthetics/e2e/journeys/synthetics/getting_started.journey.ts
@@ -6,7 +6,7 @@
  */
 
 import { journey, step, expect, before, Page } from '@elastic/synthetics';
-import { syntheticsAppPageProvider } from '../../page_objects/synthetics_app';
+import { syntheticsAppPageProvider } from '../../page_objects/synthetics/synthetics_app';
 import { cleanTestMonitors } from './services/add_monitor';
 
 journey(`Getting Started Page`, async ({ page, params }: { page: Page; params: any }) => {
