+++ x-pack/plugins/synthetics/e2e/journeys/uptime/alerts/default_email_settings.ts
@@ -19,7 +19,7 @@ import {
   byTestId,
   waitForLoadingToFinish,
 } from '@kbn/observability-plugin/e2e/utils';
-import { settingsPageProvider } from '../../page_objects/settings';
+import { settingsPageProvider } from '../../../page_objects/uptime/settings';
 
 journey('DefaultEmailSettings', async ({ page, params }) => {
   const settings = settingsPageProvider({ page, kibanaUrl: params.kibanaUrl });
