+++ x-pack/plugins/synthetics/e2e/journeys/uptime/index.ts
@@ -1,0 +1,19 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export * from './data_view_permissions';
+export * from './read_only_user';
+export * from './alerts';
+export * from './uptime.journey';
+export * from './step_duration.journey';
+export * from './monitor_details.journey';
+export * from './monitor_name.journey';
+export * from './monitor_management.journey';
+export * from './monitor_management_enablement.journey';
+export * from './monitor_details';
+export * from './locations';
+export * from './private_locations';
