+++ x-pack/plugins/synthetics/e2e/journeys/uptime/locations/index.ts
