+++ x-pack/plugins/synthetics/e2e/journeys/uptime/monitor_details/index.ts
