+++ x-pack/plugins/synthetics/e2e/journeys/uptime/monitor_details/monitor_alerts.journey.ts
@@ -8,7 +8,7 @@
 import { journey, step, expect, before, Page } from '@elastic/synthetics';
 import { noop } from 'lodash';
 import { byTestId, delay } from '@kbn/observability-plugin/e2e/utils';
-import { monitorDetailsPageProvider } from '../../page_objects/monitor_details';
+import { monitorDetailsPageProvider } from '../../../page_objects/uptime/monitor_details';
 
 const dateRangeStart = '2019-09-10T12:40:08.078Z';
 const dateRangeEnd = '2019-09-11T19:40:08.078Z';
