+++ x-pack/plugins/synthetics/e2e/journeys/uptime/monitor_details/monitor_details.journey.ts
@@ -7,7 +7,7 @@
 
 import { journey, step, before, Page } from '@elastic/synthetics';
 import { byTestId } from '@kbn/observability-plugin/e2e/utils';
-import { monitorDetailsPageProvider } from '../../page_objects/monitor_details';
+import { monitorDetailsPageProvider } from '../../../page_objects/uptime/monitor_details';
 
 const dateRangeStart = '2019-09-10T12:40:08.078Z';
 const dateRangeEnd = '2019-09-11T19:40:08.078Z';
