+++ x-pack/plugins/synthetics/e2e/journeys/uptime/private_locations/index.ts
