+++ x-pack/plugins/synthetics/e2e/journeys/uptime/read_only_user/index.ts
