+++ x-pack/plugins/synthetics/e2e/page_objects/uptime/monitor_details.tsx
