+++ x-pack/plugins/synthetics/e2e/page_objects/uptime/settings.tsx
@@ -7,8 +7,8 @@
 
 import { expect, Page } from '@elastic/synthetics';
 import { byTestId } from '@kbn/observability-plugin/e2e/utils';
-import { loginPageProvider } from './login';
-import { utilsPageProvider } from './utils';
+import { loginPageProvider } from '../login';
+import { utilsPageProvider } from '../utils';
 
 export function settingsPageProvider({ page }: { page: Page; kibanaUrl: string }) {
   return {
